Authorize TrainingController.cs

+++ ICPC WebSite-Backend/Controllers/TrainingController.cs
@@ -1,7 +1,10 @@
-﻿using ICPC_WebSite_Backend.Data.Models.DTO;
+﻿using ICPC_WebSite_Backend.Data;
+using ICPC_WebSite_Backend.Data.Models.DTO;
 using ICPC_WebSite_Backend.Data.Response;
 using ICPC_WebSite_Backend.Repository;
+using ICPC_WebSite_Backend.Security;
 using ICPC_WebSite_Backend.Utility;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -12,23 +15,52 @@ namespace ICPC_WebSite_Backend.Controllers
     public class TrainingController : ControllerBase
     {
         private readonly ITrainingRepository _trainingRepository;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly ApplicationDbContext _applicationDbContext;
 
-        public TrainingController(ITrainingRepository trainingRepository) {
+        public TrainingController(ITrainingRepository trainingRepository, IAuthorizationService authorizationService,ApplicationDbContext applicationDbContext)
+        {
             _trainingRepository = trainingRepository;
+            _authorizationService = authorizationService;
+            _applicationDbContext = applicationDbContext;
         }
+        [HttpGet]
+        public async Task<int> GetCommunityId(int trainingId)
+        {
+            var training = await _applicationDbContext.Trainings.FindAsync(trainingId);
+            return training == null ? 0 : training.Id;
+        }
+        [Authorize]
+        [HttpGet("IsAuthorized")]
+        public async Task<bool> IsAuthorized(int communityId)
+        {
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, new ClaimResource(ClaimsNames.CommunityIdClaimName, communityId), "EditAccess");
+            return authorizationResult.Succeeded;
+        }
+
+        [Authorize(Roles = RolesList.Administrator + "," + RolesList.CommunityLeader + "," + RolesList.HeadOfTraining)]
         [HttpPost("addtraining")]
-        public async Task<IActionResult> addtraining(TrainingDTO trainingDTO) {
+        public async Task<IActionResult> addtraining(TrainingDTO trainingDTO)
+        {
+            if (!await IsAuthorized(trainingDTO.Community_Id)) return Forbid();
+
             var validate = Validate.IsValidTraining(trainingDTO);
             if (!validate.Succeeded)
                 return BadRequest(validate);
             var result = await _trainingRepository.AddTraining(trainingDTO);
-            if (!result.Succeeded) {
+            if (!result.Succeeded)
+            {
                 return Unauthorized(result.Errors);
             }
+
             return Ok(result);
         }
+        [Authorize(Roles = RolesList.Administrator + "," + RolesList.CommunityLeader + "," + RolesList.HeadOfTraining)]
         [HttpPut("updatetraining")]
         public async Task<IActionResult> updatetraining(int trainingId, TrainingDTO trainingDTO) {
+            
+            if (!await IsAuthorized(await GetCommunityId(trainingId))) return Forbid();
+
             var validate = Validate.IsValidTraining(trainingDTO);
             if (!validate.Succeeded)
                 return BadRequest(validate);
@@ -54,33 +86,44 @@ namespace ICPC_WebSite_Backend.Controllers
             }
             return Ok(result);
         }
+        [Authorize(Roles = RolesList.Administrator + "," + RolesList.CommunityLeader + "," + RolesList.HeadOfTraining)]
+
         [HttpDelete("deletetraining")]
         public async Task<IActionResult> deletetraining(int trainingId) {
+            if (!await IsAuthorized(await GetCommunityId(trainingId))) return Forbid();
+
             var result = await _trainingRepository.DeleteTraining(trainingId);
             if (!result.Succeeded) {
                 return NotFound(result.Errors);
             }
             return Ok(result);
         }
-        
+        [Authorize(Roles = RolesList.Administrator + "," + RolesList.CommunityLeader + "," + RolesList.HeadOfTraining)]
+
         [HttpPost("linkweek")]
         public async Task<IActionResult> LinkWeek(int trainingId, int weekId)
         {
+            if (!await IsAuthorized(await GetCommunityId(trainingId))) return Forbid();
+
             var result = await _trainingRepository.LinkWeek(trainingId, weekId);
             if (!result.Succeeded) {
                 return NotFound(result.Errors);
             }
             return Ok(result);
         }
-       [ HttpDelete("Unlinkweek")]
+        [Authorize(Roles = RolesList.Administrator + "," + RolesList.CommunityLeader + "," + RolesList.HeadOfTraining)]
+        [HttpDelete("Unlinkweek")]
         public async Task<IActionResult> UnLinkWeek(int trainingId, int weekId)
         {
+            if (!await IsAuthorized(await GetCommunityId(trainingId))) return Forbid();
+
             var result = await _trainingRepository.UnLinkWeek(trainingId, weekId);
             if (!result.Succeeded) {
                 return NotFound(result.Errors);
             }
             return Ok(result);
         }
+        [Authorize]
         [HttpPost("Join/{trainingId}")]
         public async Task<IActionResult> JoinTraining([FromQuery] string userId, int trainingId) {
             var result = await _trainingRepository.JoinTraining(userId, trainingId);
@@ -89,20 +132,24 @@ namespace ICPC_WebSite_Backend.Controllers
             }
             return Ok(result);
         }
+        [Authorize(Roles = RolesList.Administrator + "," + RolesList.CommunityLeader + "," + RolesList.HeadOfTraining)]
         [HttpPost("RequestRespond/{trainingId}")]
         public async Task<IActionResult> ApproveJoinTrainingRequest([FromQuery] string userId, int trainingId, [FromQuery] bool approve) {
+            if (!await IsAuthorized(await GetCommunityId(trainingId))) return Forbid();
+
             var result = await _trainingRepository.ResponseToTrainingRequest(userId, trainingId, approve);
             if (!result.Succeeded) {
                 return Unauthorized(result);
             }
             return Ok(result);
         }
+        [Authorize]
         [HttpGet("Members/{trainingId}")]
         public async Task<IActionResult> GetTrainingMembers(int trainingId, [FromQuery] string status)
         {
             if (status == ConstVariable.PendingStatus)
             {
-                //check for authorization here
+                if (!await IsAuthorized(await GetCommunityId(trainingId))) return Forbid();
             }
 
             var result = await _trainingRepository.GetTrainingMembersAsync(trainingId, status);
